+++ lib/src/canvas_context/model/component_data.dart
@@ -4,7 +4,7 @@ import 'package:uuid/uuid.dart';
 
 class ComponentData with ChangeNotifier {
   /// Unique id of this component.
-  String id;
+  final String id;
 
   /// Position on the canvas.
   Offset position;
@@ -20,7 +20,7 @@ class ComponentData with ChangeNotifier {
   /// Component type to distinguish components.
   ///
   /// You can use it for example to distinguish what [data] type this component has.
-  final String type;
+  final String? type;
 
   /// This value determines if this component will be above or under other components.
   /// Higher value means on the top.
@@ -30,7 +30,7 @@ class ComponentData with ChangeNotifier {
   ///
   /// Use for hierarchical components.
   /// Functions such as [moveComponentWithChildren] work with this property.
-  String parentId;
+  String? parentId;
 
   /// List of children of this component.
   ///
@@ -49,20 +49,14 @@ class ComponentData with ChangeNotifier {
 
   /// Represents data of a component in the model.
   ComponentData({
-    this.id,
+    String? id,
     this.position = Offset.zero,
     this.size = const Size(80, 80),
     this.minSize = const Size(4, 4),
     this.type,
     this.data,
   })  : assert(minSize <= size),
-        assert(position != null),
-        assert(size != null),
-        assert(minSize != null) {
-    if (this.id == null) {
-      this.id = Uuid().v4();
-    }
-  }
+        this.id = id ?? Uuid().v4();
 
   /// Updates this component on the canvas.
   ///